(feat) landing page, theme, misc
