--- src/Components/Dashboard.tsx
@@ -1,9 +1,0 @@
-import { UserCollegeData } from "../types";
-
-interface DashboardInt {
-  data: UserCollegeData;
-}
-
-export function Dashboard({ data }: DashboardInt) {
-  return <>{JSON.stringify(data)}</>;
-}
