+++ src/Cookie.ts
@@ -54,3 +54,27 @@ export function isLoggedIn() {
   }
   return false;
 }
+
+export function getProfilePicture() {
+  const user_info = localStorage.getItem("user_info");
+  if (user_info) {
+    return JSON.parse(user_info).result.profilePicture + "";
+  }
+  return "";
+}
+
+export function getProfileName() {
+  const user_info = localStorage.getItem("user_info");
+  if (user_info) {
+    const first_name =
+      JSON.parse(user_info).result.firstName === "NA"
+        ? ""
+        : JSON.parse(user_info).result.firstName;
+    const last_name =
+      JSON.parse(user_info).result.lastName === "NA"
+        ? ""
+        : JSON.parse(user_info).result.lastName;
+    return first_name + " " + last_name;
+  }
+  return "";
+}
