+++ src/theme/styles.css
@@ -1,0 +1,1 @@
+@import url("https://fonts.googleapis.com/css2?family=Actor&family=Anaheim&family=Bakbak+One&family=Inter:wght@300;400;500;600;700;800;900&display=swap");
